rag-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pymupdf==1.24.10
openai==1.51.0
anthropic==0.34.2
psycopg2-binary==2.9.9
pgvector==0.3.4
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
tiktoken==0.7.0
numpy==1.26.4
